refactor lost password
