+++ scripts/config/pages.pl
@@ -89,12 +89,12 @@ sub all_pages {
         ['resources/market_timesws', 'resources/market_timesws',  'default', 'Trading Times'],
 
         ['user/authenticatews',              'user/authenticatews',                  'default',    'Authenticate',            'ja'],
-        ['user/lost_passwordws',             'user/lost_passwordws',                 'default',    'Password Reset'],
+        ['user/lost_passwordws',             'user/lost_password',                   'default',    'Password Reset'],
         ['user/metatrader',                  'user/metatrader',                      'full_width', 'MetaTrader account management'],
         ['user/portfoliows',                 'user/portfoliows',                     'default',    'Portfolio'],
         ['user/profit_tablews',              'user/profit_tablews',                  'default',    'Profit Table'],
         ['user/securityws',                  'user/securityws',                      'default',    'Security'],
-        ['user/settingsws',                  'user/settingsws',                      'default',    'Settings'],
+        ['user/settingsws',                  'user/settings',                        'default',    'Settings'],
         ['user/statementws',                 'user/statementws',                     'default',    'Statement'],
         ['user/tnc_approvalws',              'legal/tnc_approvalws',                 'default',    'Terms and Conditions Approval'],
         ['user/reset_passwordws',            'user/reset_passwordws',                'default',    'Password Reset'],
@@ -102,8 +102,8 @@ sub all_pages {
         ['user/reality_check_summaryws',     'user/reality_check_summaryws',         'default',    'Reality Check'],
         ['user/security/api_tokenws',        'user/security/api_tokenws',            'default',    'API Token',               'ja'],
         ['user/security/authorised_appsws',  'user/security/authorised_appsws',      'default',    'Authorised Applications', 'ja'],
-        ['user/security/cashier_passwordws', 'user/security/cashier_passwordws',     'default',    'Cashier Password'],
-        ['user/security/change_passwordws',  'user/security/change_passwordws',      'default',    'Change Password'],
+        ['user/security/cashier_passwordws', 'user/security/cashier_password',       'default',    'Cashier Password'],
+        ['user/security/change_passwordws',  'user/security/change_password',        'default',    'Change Password'],
         ['user/security/iphistoryws',        'user/security/iphistory',              'default',    'Login History',           'ja'],
         ['user/security/limitsws',           'user/security/limitsws',               'default',    'Account Limits'],
         ['user/security/self_exclusionws',   'user/security/self_exclusionws',       'default',    'Account Details'],
