+++ src/javascript/binary/websocket_pages/user/lost_password.js
@@ -1,15 +1,32 @@
-const LostPassword = require('./lost_password/lost_password.init').LostPassword;
-const Client       = require('../../base/client').Client;
+const Client         = require('../../base/client').Client;
+const localize       = require('../../base/localize').localize;
+const url_for        = require('../../base/url').url_for;
+const FormManager    = require('../../common_functions/form_manager');
 
-const LostPasswordWS = (function() {
-    const onLoad = function() {
-        if (Client.redirect_if_login()) {
-            return;
+const LostPassword = (function() {
+    'use strict';
+
+    const responseHandler = function(response) {
+        if (response.verify_email) {
+            window.location.href = url_for('user/reset_passwordws');
+        } else if (response.error) {
+            $('#form_error').removeClass('invisible').text(localize(response.error.message));
         }
-        BinarySocket.init({
-            onmessage: LostPassword.lostPasswordWSHandler,
-        });
-        LostPassword.init();
+    };
+
+    const init = function() {
+        const form_id = '#frm_lost_password';
+        FormManager.init(form_id, [
+            { selector: '#email', validations: ['req', 'email'], request_field: 'verify_email' },
+            { request_field: 'type', value: 'reset_password' },
+        ]);
+        FormManager.handleSubmit(form_id, {}, responseHandler);
+    };
+
+    const onLoad = function() {
+        if (!Client.redirect_if_login()) {
+            init();
+        }
     };
 
     return {
@@ -17,6 +34,4 @@ const LostPasswordWS = (function() {
     };
 })();
 
-module.exports = {
-    LostPasswordWS: LostPasswordWS,
-};
+module.exports = LostPassword;
