+++ src/templates/user/lost_password.html.tt
@@ -1,0 +1,28 @@
+<div class="static_full" id="lost_passwordws">
+    <h1>[% l("Password reset") %]</h1>
+    <p>[% l("To reset your password, enter the e-mail address you registered in the field below and click 'Reset Password'.") %]</p>
+
+    <form class="frm_lost_password">
+        <fieldset>
+            <div class="gr-row gr-padding-10">
+                <div class="gr-4 form_label">
+                    <label for="email">[% l("Email address") %]</label>
+                </div>
+                <div class="gr-8">
+                    <input id="email" name="email" type="text" />
+                </div>
+            </div>
+
+            <div class="gr-row">
+                <div class="gr-4"></div>
+                    <div class="gr-8">
+                        <p id="form_error" class="errorfield hidden"></p>
+                        <button class="button" type="button">[% l("Reset Password") %]</button>
+                    </div>
+                </div>
+        </fieldset>
+    </form>
+
+    <p>[% l("We'll then send you an e-mail asking you to verify your identity.") %]</p>
+    <p>[% l("If you don't receive the email within the next few minutes, please check your junk/spam mail folder.") %]</p>
+</div>
