--- src/templates/user/lost_passwordws.html.tt
@@ -1,41 +1,0 @@
-<div class="gr-12 static_full invisible" id="lost_passwordws">
-  <h1>
-    [% l("Password reset") %]
-  </h1>
-  <p>
-    [% l("To reset your password, enter the e-mail address you registered in the field below and click 'Reset Password'.") %]
-  </p>
-  <div>
-    <fieldset>
-      <div class="gr-row gr-padding-10">
-        <div class="gr-4 form_label">
-          <label for="lp_email">
-            [% l("Email address") %]
-          </label>
-        </div>
-        <div class="gr-8">
-          <input id="lp_email" name="lp_email" type="text"/>
-          <p class="errorfield invisible" id="email_error">
-            [% l("This field is required.") %]
-          </p>
-        </div>
-      </div>
-      <div class="gr-row">
-        <div class="gr-4"></div>
-        <div class="gr-8">
-          <span class="button">
-            <button class="button" id="submit" name="submit" type="button">
-              [% l("Reset Password") %]
-            </button>
-          </span>
-        </div>
-      </div>
-    </fieldset>
-  </div>
-  <p>
-    [% l("We'll then send you an e-mail asking you to verify your identity.") %]
-  </p>
-  <p>
-    [% l("If you don't receive the email within the next few minutes, please check your junk/spam mail folder.") %]
-  </p>
-</div>
