+++ src/templates/user/security/cashier_password.html.tt
