+++ src/templates/user/security/change_password.html.tt
